+++ lib/rubocop/cop/norb_cops.rb
@@ -2,5 +2,6 @@
 
 require_relative 'mixin/target_rails_version'
 
+require_relative 'norb/active_record_through_business'
 require_relative 'norb/misplaced_branching_logic'
 require_relative 'norb/namespaced_active_record'