+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 * Adding `Norb/MisplacedBranchingLogic` cop to ensure branching logic is not in the wrong place. ([@tleish][])
 * Adding `Norb/NamespacedActiveRecord` cop to ensure ActiveRecord models are namespaced. ([@tleish][])
+* Adding `Norb/ActiveRecordThroughBusiness` cop to ensure direct ActiveRecord calls come from business objects. ([@tleish][])
 
 
 [@tleish]: https://github.com/tleish